Properties extracted to common

+++ cmd/common/communication/messages.go
@@ -5,7 +5,7 @@ type FullReview struct {
 	ReviewId 		string 					`json:"review_id",omitempty`
 	UserId 			string 					`json:"user_id",omitempty`
 	BusinessId 		string 					`json:"business_id",omitempty`
-	Stars 			int 					`json:"stars",omitempty`
+	Stars 			float32 				`json:"stars",omitempty`
 	Useful			int 					`json:"useful",omitempty`
 	Funny 			int 					`json:"funny",omitempty`
 	Cool			int 					`json:"cool",omitempty`
@@ -73,5 +73,5 @@ type DishashData struct {
 // Best-Users flow.
 type StarsData struct {
 	UserId 			string 					`json:"user_id",omitempty`
-	Stars			int 					`json:"stars",omitempty`
+	Stars			float32 				`json:"stars",omitempty`
 }
